pr_review/requirements.txt
requests>=2.31.0
PyGithub>=1.59.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
typing-extensions>=4.8.0
radon>=6.0.1 